requirements.txt
boto3>=1.34.0
botocore>=1.34.0
pandas>=2.2.0
matplotlib>=3.8.0
reportlab>=4.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.50.0
vertexai>=1.50.0
Pillow>=10.0.0
